Refactor, Hibernate update, H2 database

+++ src/test/java/com/java/ds/expenditure/model/payments/PaymentTest.java
@@ -11,24 +11,24 @@ public class PaymentTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void testPaymentTitleIsNotEmpty(){
-        payment = new Payment("", BigDecimal.valueOf(100.00), LocalDate.of(2021, 12, 10),1 );
+        payment = new Payment("", 100.0, LocalDate.of(2021, 12, 10),1 );
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testSetEmptyPaymentTitle(){
-        payment = new Payment("Online payment", BigDecimal.valueOf(100.00), LocalDate.of(2021, 12, 10),1);
+        payment = new Payment("Online payment", 100.0, LocalDate.of(2021, 12, 10),1);
         payment.setTitle("");
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testPaymentValueIsPositive(){
-        payment = new Payment("Online payment", BigDecimal.valueOf(-100.0), LocalDate.of(2021, 12, 10),1);
+        payment = new Payment("Online payment", -100.0, LocalDate.of(2021, 12, 10),1);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testSetNegativePaymentValue(){
-        payment = new Payment("Online payment", BigDecimal.valueOf(100.0), LocalDate.of(2021, 12, 10),1);
-        payment.setValue(BigDecimal.valueOf(-100.0));
+        payment = new Payment("Online payment", 100.0, LocalDate.of(2021, 12, 10),1);
+        payment.setValue(-100.0);
     }
 
 }
